Some error_message refactoring

+++ code/MapData.hpp
@@ -168,7 +168,7 @@ private:
 class Field {
 public:
 
-    Field(const int &width, const int &height,
+    Field(const double &width, const double &height,
           const std::vector<Area> &walls,
           const std::vector<Area> &traps,
           const std::vector<Area> &swamps,
